backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-static==0.0.6
Django==4.2.20
django-cors-headers==4.4.0
django-database-url==1.0.3
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
static3==0.7.0
typing_extensions==4.12.2
urllib3==2.2.3
